Introduce onTick() function
Instead of trying to figure out what objects to update and when and in which order, there's now onTick() function for all objects that is called on every object in the main array. Each class can do whatever it wants with it and do the proper update. This fixes few timing things and removes some special signals.